+++ unitTests/Test Cases/SocializeErrorTests.h
@@ -1,0 +1,13 @@
+//
+//  SocializeErrorTests.h
+//  SocializeSDK
+//
+//  Created by Nathaniel Griswold on 12/3/11.
+//  Copyright (c) 2011 Socialize, Inc. All rights reserved.
+//
+
+#import <GHUnitIOS/GHUnit.h>
+
+@interface SocializeErrorTests : GHTestCase
+
+@end